+++ springbox-catalog/src/main/java/io/springbox/catalog/repositories/GenreRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 
 import io.springbox.catalog.domain.Genre;
 
-public interface GenreRepository extends CrudRepository<Genre, Long> {
+public interface GenreRepository extends CrudRepository<Genre, Integer> {
 }
